Set up webpack
