+++ src/js/players.js
@@ -2,3 +2,5 @@
 function Player(name, sign) {
   return { name, sign };
 }
+
+export { Player as default };
