+++ webpack.config.js
@@ -1,0 +1,20 @@
+const path = require('path');
+
+module.exports = {
+  entry: './src/js/DOM.js',
+  output: {
+    filename: 'main.js',
+    path: path.resolve(__dirname, 'dist'),
+  },
+  module: {
+    rules: [
+      {
+        test: /\.css$/,
+        use: [
+          'style-loader',
+          'css-loader',
+        ],
+      },
+    ],
+  },
+};
